// app/backend/src/interfaces/Filter.ts
export default interface IntFilterTeam {
  homeTeamGoals: number;
  awayTeamGoals: number;
}

// export default interface IntLeaderboard {
//   name: string;
//   totalPoints: number;
//   totalGames: number;
//   totalVictories: number;
//   totalDraws: number;
//   totalLosses: number;
//   goalsOwn: number;
//   goalsFavor: number;
//   goalsBalance: number;
//   efficiency: number;
// }